penambahan midlleware dan gate

+++ app/Http/Controllers/TodoController.php
@@ -6,6 +6,7 @@ use App\Models\Todo;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class TodoController extends Controller
@@ -16,8 +17,8 @@ class TodoController extends Controller
 
         return view('welcome', [
             'todos' =>Todo::all(),
-            'todo' => null,
-            'todoUser' => auth()->user() ? $user->find(auth()->user()->id)->todos: null
+            'todo' => NULL,
+            'todoUser' => auth()->user() ? $user->find(auth()->user()->id)->todos: NULL
         ]);
     }
 
@@ -44,13 +45,21 @@ class TodoController extends Controller
 
     public function checked($id) : RedirectResponse {
         $todo = Todo::findOrFail($id);
+        $todo = Todo::findOrFail($id);
+        if (!Gate::allows('update-todo', $todo)) {
+             abort(403, 'unauthorized');
+        };
         $todo->isDone = true;
         $todo->save();
          return redirect('/');
     }
 
     public function update($id, Request $request) : RedirectResponse {
+        
         $todo = Todo::findOrFail($id);
+        if (!Gate::allows('update-todo', $todo)) {
+             abort(403, 'unauthorized');
+        };
         $validated = $request->validate([
             'activity' => 'required|min:4',
         ]);
@@ -61,15 +70,25 @@ class TodoController extends Controller
 
     public function deleteTodo($id) : RedirectResponse {
         $todo = Todo::findOrFail($id);
+        $todo = Todo::findOrFail($id);
+        if (!Gate::allows('update-todo', $todo)) {
+             abort(403, 'unauthorized');
+        };
         $todo->delete();
          return redirect('/');
     }
 
     public function showData($id) : View {
+        $user = new User();
         $todo = Todo::findOrFail($id);
+        $todo = Todo::findOrFail($id);
+        if (!Gate::allows('update-todo', $todo)) {
+             abort(403, 'unauthorized');
+        };
         return view('welcome', [
             'todo' => $todo,
-            'todos' =>Todo::all()
+            'todoUser' => auth()->user() ? $user->find(auth()->user()->id)->todos: NULL
+        
         ] );
     }
 }
